# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Business OS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/services": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "List Services",
        "operationId": "list_services",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  },
                  "title": "Response List Services"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "services"
        ],
        "summary": "Create Service",
        "operationId": "create_service",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/services/{service_id}": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "Get Service",
        "operationId": "get_service",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "services"
        ],
        "summary": "Update Service",
        "operationId": "update_service",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/modules/catalog": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Catalog",
        "operationId": "catalog",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{3}$",
              "default": "ILS",
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/modules/recommend": {
      "post": {
        "tags": [
          "modules"
        ],
        "summary": "Recommend Modules",
        "operationId": "recommend_modules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionnaireIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recommendation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/modules/quote": {
      "post": {
        "tags": [
          "modules"
        ],
        "summary": "Quote Modules",
        "operationId": "quote_modules",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{3}$",
              "default": "ILS",
              "title": "Currency"
            }
          },
          {
            "name": "active_clients",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Active Clients"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Selection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/current/modules": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Get Tenant Modules",
        "operationId": "get_tenant_modules",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantModules"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "modules"
        ],
        "summary": "Put Tenant Modules",
        "description": "Replaces the business's modules (billing is simulated in the prototype).",
        "operationId": "put_tenant_modules",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Selection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantModules"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Catalog": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "core": {
            "items": {
              "$ref": "#/components/schemas/CoreTier"
            },
            "type": "array",
            "title": "Core"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/CatalogModule"
            },
            "type": "array",
            "title": "Modules"
          },
          "presets": {
            "additionalProperties": {
              "items": {
                "type": "string",
                "enum": [
                  "client_app",
                  "ai_basic",
                  "ai_pro",
                  "crm",
                  "analytics_pro",
                  "agent_finance",
                  "agent_marketing",
                  "whatsapp",
                  "extra_location"
                ]
              },
              "type": "array"
            },
            "propertyNames": {
              "enum": [
                "starter",
                "growing",
                "ai_powered"
              ]
            },
            "type": "object",
            "title": "Presets"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "core",
          "modules",
          "presets"
        ],
        "title": "Catalog"
      },
...
      "CoreTier": {
        "properties": {
          "up_to_clients": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Up To Clients"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "up_to_clients",
          "price"
        ],
        "title": "CoreTier"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionnaireIn": {
        "properties": {
          "active_clients": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Active Clients"
          },
          "staff": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Staff"
          },
          "locations": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Locations"
          },
          "wants_client_app": {
            "type": "boolean",
            "title": "Wants Client App"
          },
          "wants_ai_actions": {
            "type": "boolean",
            "title": "Wants Ai Actions"
          }
        },
        "type": "object",
        "required": [
          "active_clients",
          "staff",
          "locations",
          "wants_client_app",
          "wants_ai_actions"
        ],
        "title": "QuestionnaireIn"
      },
      "QuoteOut": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "core": {
            "type": "integer",
            "title": "Core"
          },
          "core_tier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Tier"
          },
          "lines": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "client_app",
                "ai_basic",
                "ai_pro",
                "crm",
                "analytics_pro",
                "agent_finance",
                "agent_marketing",
                "whatsapp",
                "extra_location"
              ]
            },
            "type": "object",
            "title": "Lines"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "core",
          "core_tier",
          "lines",
          "total"
        ],
        "title": "QuoteOut"
      },
...
      "Recommendation": {
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "starter",
              "growing",
              "ai_powered"
            ],
            "title": "Preset"
          },
          "modules": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "client_app",
                "ai_basic",
                "ai_pro",
                "crm",
                "analytics_pro",
                "agent_finance",
                "agent_marketing",
                "whatsapp",
                "extra_location"
              ]
            },
            "type": "object",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "preset",
          "modules"
        ],
        "title": "Recommendation"
      },
...
      "Selection": {
        "properties": {
          "modules": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "client_app",
                "ai_basic",
                "ai_pro",
                "crm",
                "analytics_pro",
                "agent_finance",
                "agent_marketing",
                "whatsapp",
                "extra_location"
              ]
            },
            "type": "object",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "modules"
        ],
        "title": "Selection"
      },
...
      "Service": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "price_amount": {
            "type": "integer",
            "title": "Price Amount"
          },
          "price_currency": {
            "type": "string",
            "title": "Price Currency"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "duration_minutes",
          "capacity",
          "price_amount",
          "price_currency",
          "color",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "Service"
      },
      "ServiceCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "duration_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Duration Minutes"
          },
          "capacity": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Capacity",
            "default": 1
          },
          "price_amount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Price Amount",
            "description": "Price in minor units (agorot, cents)",
            "default": 0
          },
          "price_currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Currency",
            "description": "Defaults to the business currency"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "duration_minutes"
        ],
        "title": "ServiceCreate"
      },
...
      "ServiceUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "price_amount": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Amount"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "ServiceUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}